Wave 5: Decompose pages/index.tsx and finalize strict TypeScript
- Extract useSubscriptionFilters: filter/sort state and memoized derived list
- Extract useSubscriptionMutations: pending-change state, dialog refs, CRUD handlers
- Extract SummaryHeader component: totals hero + credit card rail (with scoped CSS)
- Extract SubscriptionList component: subscription cards, add placeholder, confirm dialogs
- Convert firestoreSeed.mjs → firestoreSeed.ts; run via tsx; add tsx devDep
- Convert helpers/helpers.test.js → .ts; add @types/jest; type test fixtures
- Flip allowJs: false; drop **/*.js/**/*.jsx from tsconfig include
- pages/index.tsx reduced from 678 → 273 lines (nav layout + CSS only)
- pnpm tsc --noEmit: clean; pnpm test: 4/4 passing
