+++ package.json
@@ -7,7 +7,7 @@
     "node": ">=20 <25"
   },
   "scripts": {
-    "seed": "node scripts/firebase/firestoreSeed.mjs",
+    "seed": "tsx scripts/firebase/firestoreSeed.ts",
     "dev": "next dev",
     "debug": "NODE_OPTIONS='--inspect' next dev",
     "build": "next build",
@@ -37,6 +37,7 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^14.3.1",
+    "@types/jest": "^30.0.0",
     "@types/node": "^20.17.0",
     "@types/react": "^18.3.0",
     "@types/react-dom": "^18.3.0",
@@ -48,6 +49,7 @@
     "jest-environment-jsdom": "^29.7.0",
     "lint-staged": "^15.2.0",
     "prettier": "^3.3.0",
+    "tsx": "^4.19.0",
     "typescript": "^5.6.0"
   }
 }
